Initial commit-  PySide6 Hermite Spline class
Qt only has Bezier splines, which are weird for curve editing. This implements multipoint Hermite Cubics, with "tangent" handles. Includes a test rig, so this is a fully runnable fragment - will become a gist at some point.
There are potential arguments about the drawing of the tangents vs the pure maths underlying them, but this will work for editing use-cases, and can be extended to a- and nonsymmetric tangents quite easily.